Clean up the code.
Since now the design of Pulsed has significant changes from previous,
clean up the unused code.
